test(workflow-core): consolidate into existing CoreExecutorReflectionSpec
Address Copilot review on #5400: CoreExecutorReflectionSpec already
covered OperatorExecutor / SourceOperatorExecutor defaults and
ExecFactory.newExecFromJavaClassName / JavaRuntimeCompilation. I
missed the existing spec during my initial coverage check (its name
doesn't match any source-class name pattern). Drop the duplicate
CoreExecutorSpec.scala and instead add five genuinely-new tests to
the existing spec:

- OperatorExecutor.open / close honor a subclass override across the
  lifecycle (counter-based pin so direct no-op stubs would not pass).
- processTupleMultiPort forwards the input port id verbatim to the
  subclass override (regression that hard-codes port = 0 would be
  invisible to the basic delegation test).
- processTupleMultiPort produces as many output pairs as
  processTuple emits (zero / many fan-out).
- SourceOperatorExecutor.processTupleMultiPort never invokes
  produceTuple on the input-side path (counter-based pin — an
  accidental drain that discards results wouldn't fail an
  empty-iterator check).
- ExecFactory.newExecFromJavaClassName propagates
  NoSuchMethodException when neither factory branch matches a
  fixture with only a (Long) constructor.

The PR diff is now additive on top of the existing spec rather than
a parallel duplicate.
